requirements.txt
Django~=4.1.3
Pillow~=9.3.0
django-crispy-forms
django-taggit